@@ -1,23 +1,17 @@
 package be.jago.websocket.config;
 
-import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.CacheControl;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
-import org.springframework.web.client.RestTemplate;
-
-import java.time.Duration;
+import org.springframework.web.client.RestClient;
 
 @Configuration
 public class ServerConfig {
     @Bean
-    public RestTemplate restTemplate(RestTemplateBuilder builder) {
-        return builder
-                .setConnectTimeout(Duration.ofMillis(10000))
-                .setReadTimeout(Duration.ofMillis(10000))
-                .build();
+    public RestClient restClient() {
+        return RestClient.create();
     }
 
     @Bean
