@@ -1,40 +1,40 @@
-var stompClient = null;
+    var stompClient = null;
 
-function setConnected(connected) {
-    $("#connect").prop("disabled", connected);
-    $("#disconnect").prop("disabled", !connected);
-    if (connected) {
-        $("#jokelist").show();
+    function setConnected(connected) {
+        $("#connect").prop("disabled", connected);
+        $("#disconnect").prop("disabled", !connected);
+        if (connected) {
+            $("#jokelist").show();
+        }
     }
-}
 
-function connect() {
-    var socket = new SockJS('/websocket');
-    stompClient = Stomp.over(socket);
-    stompClient.connect({}, function () {
-        setConnected(true);
-        stompClient.subscribe('/joke/jokestream', function (joke) {
-            showJoke(JSON.parse(joke.body).value);
+    function connect() {
+        var socket = new SockJS('/websocket');
+        stompClient = Stomp.over(socket);
+        stompClient.connect({}, function () {
+            setConnected(true);
+            stompClient.subscribe('/joke/joke-stream', function (joke) {
+                showJoke(JSON.parse(joke.body).value);
+            });
         });
-    });
-}
-
-function disconnect() {
-    if (stompClient !== null) {
-        stompClient.disconnect();
     }
-    setConnected(false);
-}
 
-function showJoke(joke) {
-    $("#jokes").append("<tr><td>" + joke + "</td></tr>");
-}
+    function disconnect() {
+        if (stompClient !== null) {
+            stompClient.disconnect();
+        }
+        setConnected(false);
+    }
 
-$(function () {
-    $("form").on('submit', function (e) {
-        e.preventDefault();
+    function showJoke(joke) {
+        $("#jokes").append("<tr><td>" + joke + "</td></tr>");
+    }
+
+    $(function () {
+        $("form").on('submit', function (e) {
+            e.preventDefault();
+        });
+        $( "#connect" ).click(function() { connect(); });
+        $( "#disconnect" ).click(function() { disconnect(); });
     });
-    $( "#connect" ).click(function() { connect(); });
-    $( "#disconnect" ).click(function() { disconnect(); });
-});
 
